requirements.txt
# ============================================================
# REASONING CORE — Digital Human OS
# requirements.txt
# Python 3.11+ required
# ============================================================

# ── CORE GRAPH ORCHESTRATION ─────────────────────────────────
langgraph>=0.2.55
langgraph-checkpoint-postgres>=2.0.9
langchain-core>=0.3.29

# ── LLM PROVIDERS ────────────────────────────────────────────
anthropic>=0.45.0
langchain-anthropic>=0.3.7
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0

# ── MEMORY ───────────────────────────────────────────────────
chromadb>=0.6.3
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.36

# ── TOOLS ────────────────────────────────────────────────────
tavily-python>=0.5.0
httpx>=0.28.1

# ── WEB UI (Phase 2) ─────────────────────────────────────────
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
websockets>=13.0
jinja2>=3.1.4

# ── VOICE (Phase 2) ──────────────────────────────────────────
SpeechRecognition>=3.10.0
deepgram-sdk>=3.0.0
elevenlabs>=1.0.0
pyaudio>=0.2.14

# ── OBSERVABILITY ────────────────────────────────────────────
langsmith>=0.2.11

# ── CONFIG & ENV ─────────────────────────────────────────────
python-dotenv>=1.0.1
pydantic>=2.10.6
pydantic-settings>=2.7.0

# ── ASYNC ────────────────────────────────────────────────────
anyio>=4.7.0

# ── UTILITIES ────────────────────────────────────────────────
psutil>=6.1.0
rich>=13.9.4
tenacity>=9.0.0
tiktoken>=0.8.0
structlog>=24.4.0

# ── TESTING ──────────────────────────────────────────────────
pytest>=8.3.4
pytest-asyncio>=0.24.0
pytest-env>=1.1.5
